requirements.txt
langchain
langchain-openai
openai
python-pptx
gitpython
python-dotenv
requests
diagrams
